Relax isolation in the container for ISO build
We need to allow access to /dev/loop* devices to 'losetup' so that
lorax can create the ISO rootfs. This would normally be done via the
DeviceAllow systemd directive. However, the nspawn version from
systemd 219 does not have the --property or --settings options and we
don't want to alter the host system by adding .slice files. Writing
directly to sysfs is not feasible because mock randomizes the
container ID and the user might have other containers as well.

Since we already use --share-system for the packages builds, lets use
it to solve this issue as well.

See:
man systemd-nspawn
man systemd.directives
man systemd.slice